adding apply Ui page as a component
