+++ src/pages/Apply/apply.module.css
@@ -1,5 +1,5 @@
 .formContainer {
-  width: 40%;
+  width: 90%;
   margin: auto;
   padding: 1rem;
   border: 1px solid #ddd;
@@ -33,37 +33,18 @@
 .submitButton:hover {
   background-color: #005bb5;
 }
-.applypage {
+.applypagefirst {
   display: flex;
   align-items: center;
+  justify-content: end;
   padding: 1rem;
+  width: 80%;
   font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
   color: #333;
 }
-
-.jobInfo {
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  padding: 1rem 1.5rem;
-  width: 50%;
-  height: auto;
-  background-color: #fafafa;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.05);
-  margin-bottom: 2rem;
-}
-
-.jobInfo h1 {
-  margin-bottom: 0.75rem;
-  font-weight: 700;
-  color: #0070f3;
-}
-
-.jobInfo p {
-  margin: 0.3rem 0;
-  font-size: 1rem;
-  line-height: 1.4;
-}
-
-.jobInfo strong {
-  color: #222;
+.applypage {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 100px;
 }
